add overlapper cloze deletion buttons
Monkeypatching the cloze buttons no longer works, as the
implementation has been moved to JavaScript, with no known way of
patching its behavior. This change adds buttons that use the same
icons and shortcuts and work both for the overlapper template and
the normal cloze template. This could have consequences, because
the new behavior emulates the normal cloze behavior rather than
delegating to the JavaScript implementation.